Struct nuclear::diagnostics::NeutronDiagnostics

pub struct NeutronDiagnostics {

    pub neutron_generation_history: Vec<i64>,
    pub neutron_position_bins: Vec<BinData>,
    pub neutron_fission_locations: Vec<Vec3D>,
    pub bin_parameters: GeometryDiagnostics,
    pub estimate_k: bool,
    pub track_bins: bool,
    pub track_fission_positions: bool,
    pub track_from_generation: i64,
    pub max_generation_value: i64,
    pub averaged_k: f64,
    pub halt_cause: SimulationHaltCauses,
    pub initial_neutron_count: i64,
    pub total_neutrons_tracked: i64,
    pub total_fissions: i64,
    pub power_generated: f64,

}

Fields
neutron_generation_history: Vec<i64>neutron_position_bins: Vec<BinData>neutron_fission_locations: Vec<Vec3D>bin_parameters: GeometryDiagnosticsestimate_k: booltrack_bins: booltrack_fission_positions: booltrack_from_generation: i64max_generation_value: i64averaged_k: f64halt_cause: SimulationHaltCausesinitial_neutron_count: i64total_neutrons_tracked: i64total_fissions: i64power_generated: f64
Implementations

impl NeutronDiagnostics

pub fn estimate_k(&mut self) -> Option<(f64, Vec<i64>)>

Estimates the multiplication coefficient k based on data collected in the neutron scheduler during a simulation run. At the end of the simulation, this vector is passed from the scheduler to diagnostics via track_simulation_halt.

impl NeutronDiagnostics

pub fn write_simulation_report( &mut self, dir_path: &Path, simulation_duration: Duration, halt_time: Option<f64> )

pub fn write_data( &mut self, simulation_duration: Duration, halt_time: Option<f64>, write_results: bool )

impl NeutronDiagnostics

pub fn new( estimate_k: bool, track_bins: bool, track_fission_positions: bool, track_from_generation: i64, bin_parameters: GeometryDiagnostics, initial_neutron_count: i64 ) -> NeutronDiagnostics

pub fn get_total_fissions(&self) -> i64

pub fn get_current_bin(&self, neutron_position: Vec3D) -> Option<usize>

pub fn track_neutron_location_fission( &mut self, generation_number: i64, neutron_position: Vec3D )

pub fn track_neutron_bin_presence( &mut self, generation_number: i64, neutron_position: Vec3D )

pub fn track_neutron_bin_fission( &mut self, generation_number: i64, neutron_position: Vec3D )

pub fn track_simulation_halt( &mut self, neutron_generation: i64, neutron_generation_history: Vec<i64>, halt_cause: SimulationHaltCauses )