Struct nuclear::geometry::components::Components

pub struct Components {
    pub material_data_vector: Vec<MaterialData>,
    pub cached_material_properties: Vec<MaterialProperties>,
    pub parts_vector: Vec<PartTypes>,
    pub cache_initialized: bool,
    pub simulation_range_squared: f64,
}

Struct that contains the material data, cached properties and parts - essentially all the geometry.

Fields
material_data_vector: Vec<MaterialData>cached_material_properties: Vec<MaterialProperties>parts_vector: Vec<PartTypes>cache_initialized: boolsimulation_range_squared: f64
Implementations

impl Components

pub fn new( material_data_vector: Vec<MaterialData>, parts_vector: Vec<PartTypes> ) -> Self

pub fn get_maximum_radius_squared(&mut self)

Automatically calculates the maximum radius squared, beyond which the neutron is discarded. If this wasn’t set correctly manually, it would mess up (if it’s too small, part of the geometry would be ignored; too large, and the simulation becomes very slow if neutrons have to escape first). The code iterates over all the parts, skipping those with order <= -1, and determines the largest bounds. This function should be ran after creation of the simulation to set components.simulation_range_squared.

pub fn update_cache_properties(&mut self, neutron_energy: f64)

Updating the cache of material properties for the given neutron’s energy. This should be done any time the neutron’s energy changes significantly, or whenever the simulation starts.

pub fn get_composition_total_cross_section( &self, part_composition_vector: &Vec<PartComposition> ) -> f64

Calculates the composition’s total cross-section for a given neutron’s energy from the cached properties. The total cross-section can then be used in the neutron dynamics to calculate interaction probabilities. If the cache is not updated, this will mess up.

pub fn check_material_fractions_sum(&self)

Sums each of the part’s material composition vectors and ensures the fractions add up to 1. If not, it throws an error. This is probably a very error-prone operation, if we specify the parts manually. If we move to configuration files later, we can move this to integration tests.

pub fn select_material_from_composition( &self, rng: &mut SmallRng, part_composition_vector: &Vec<PartComposition>, composition_total_cross_section: f64 ) -> usize

Determines which material from a given composition interacts with the neutron. This is mainly relevant for mixed materials, such as water (H-1/O-16), U-235/U-238 etc.

pub fn get_material_index( &self, rng: &mut SmallRng, neutron_position: &Vec3D ) -> (usize, f64)

Gets the material index based on the neutron’s current position by checking each individual part and their order. The part with the highest order is selected, through constructive solid geometry. The actual isotope that is interacted with is then selected from a composite material and returned, together with the total cross-section.

pub fn get_material_properties( &self, rng: &mut SmallRng, neutron_position: &Vec3D ) -> (&MaterialProperties, f64)

Gets the material properties and total cross-section based on the neutron’s current position. This requires the cache to have been updated beforehand. The function will throw an exception if this has not been done.