Struct nuclear::utils::config_loading::Config

pub struct Config {
    pub parallelization_parameters: ParallelizationParametersTOML,
    pub simulation_parameters: SimulationParametersTOML,
    pub heat_diffusion_parameters: HeatDiffusionParametersTOML,
    pub bin_parameters: BinParametersTOML,
    pub plot_parameters: PlotParametersTOML,
}

Struct containing the three parameters we can load in. Used by serde and returned to the main function.

Fields
parallelization_parameters: ParallelizationParametersTOMLsimulation_parameters: SimulationParametersTOMLheat_diffusion_parameters: HeatDiffusionParametersTOMLbin_parameters: BinParametersTOMLplot_parameters: PlotParametersTOML