Module nuclear::utils::data_handling

Interpreting data with binary search and interpolation.

Structs

BinarySearchResult

Struct used for binary search. Contains information on the value obtained and the convergence.

Functions

binary_search

Performs binary search based on an input vector and a target value, and returns an instance of BinarySearchResult.

get_watt_parameters

Returns Watt parameters through linear interpolation. Finds the index of the a parameter and uses that to also get the b parameter without running linear interpolation twice.

interpolate_values

Interpolates values for the linear interpolation function.

linear_interpolation

Linearly interpolates two vectors using binary search.