Module nuclear::utils::data_writing

Writing various data to file.

Functions

write_bin_results_grid

Write the bin results to a structured grid, with x,y,z pairs. Difficult to re-serialize but can be loaded into ParaView VTKs.

write_bin_results_vector

Writing the bin results as a vector, the same as in memory. Used for internal storage: can be directly deserialized and loaded in again.